requirements.txt
PyYAML==6.0.2
pg8000==1.31.2
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.6
requests==2.32.4